Moves IndexWriter into .gbptree and renames to Writer
