+++ community/index/src/main/java/org/neo4j/index/gbptree/ValueMerger.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.index;
+package org.neo4j.index.gbptree;
 
 /**
  * Decides what to do when inserting key which already exists in index. Different implementations of
