+++ community/index/src/main/java/org/neo4j/index/gbptree/Writer.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.index;
+package org.neo4j.index.gbptree;
 
 import java.io.Closeable;
 import java.io.IOException;
@@ -30,7 +30,7 @@ import java.io.IOException;
  * @param <KEY> type of keys
  * @param <VALUE> type of values
  */
-public interface IndexWriter<KEY,VALUE> extends Closeable
+public interface Writer<KEY,VALUE> extends Closeable
 {
     /**
      * Associate given {@code key} with given {@code value}.
